# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bistro Bliss API",
        "description": "A brief description of your API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/users": {
            "get": {
                "tags": [
                    "Admins"
                ],
                "summary": "Get all users",
                "operationId": "c457726701591d1183b53aa71fc13441",
                "responses": {
                    "200": {
                        "description": "A list of users",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {}
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/admin/create-user": {
            "post": {
                "tags": [
                    "Admins"
                ],
                "summary": "Create a new user",
                "operationId": "20dcc92a689aa2a527df22961fa2db64",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "username": {
                                        "description": "Username of the user",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Email address of the user",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Password for the user",
                                        "type": "string"
                                    },
                                    "role": {
                                        "description": "Role of the user",
                                        "type": "string",
                                        "enum": [
                                            "user",
                                            "admin"
                                        ]
                                    },
                                    "phone": {
                                        "description": "Phone number of the user",
                                        "type": "string"
                                    },
                                    "address": {
                                        "description": "Address of the user",
                                        "type": "string"
                                    },
                                    "image": {
                                        "description": "Profile image",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/user/update": {
            "patch": {
                "tags": [
                    "Users"
                ],
                "summary": "Update authenticated user's details",
                "operationId": "e6895b1c1b66782101e0fd8f07c09582",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User updated successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                }
            }
        },
        "/api/admin/update-user/{id}": {
            "put": {
                "tags": [
                    "Admins"
                ],
                "summary": "Update a user's details",
                "operationId": "f5c4e65d132fdb3aa3922aa925064748",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User updated successfully"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/admin/delete-user/{id}": {
            "delete": {
                "tags": [
                    "Admins"
                ],
                "summary": "Delete a user",
                "operationId": "67b84dee7d9da67ac0758df9c6febb14",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "User deleted successfully"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "User": {
                "title": "User",
                "description": "User object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "username": {
                        "type": "string",
                        "example": "john_doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john@example.com"
                    },
                    "phone": {
                        "type": "string",
                        "example": "+1234567890"
                    },
                    "address": {
                        "type": "string",
                        "example": "123 Main St"
                    },
                    "image": {
                        "type": "string",
                        "example": "https://example.com/image.jpg"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-10-09T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-10-09T12:00:00Z"
                    }
                },
                "type": "object"
            }
        }
    },
...
}